casting to int instead of double

+++ 0x00-hello-dart/6-positive_or_negative.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 void main(List<String> args) {
-    var arg = double.parse(args[0]);
+    var arg = int.parse(args[0]);
     stdout.write('${arg} ');
     if(arg > 0) { print('is positive'); }
     else if(arg == 0) { print('is zero'); }
